validation

+++ app/Http/Controllers/Country/CountryController.php
@@ -9,7 +9,11 @@ namespace App\Http\Controllers\Country;
 
 use App\Http\Controllers\Controller;
 use App\Models\CountryModel;
+use Dotenv\Validator as DotenvValidator;
 use Illuminate\Http\Request;
+use App\Http\Requests\CountryValidator;
+
+use Validator;
 
 class CountryController extends Controller
 {
@@ -18,10 +22,32 @@ class CountryController extends Controller
     }
 
     public function countryById($id){
-        return response()->json(CountryModel::find($id), 200);
+        $country = CountryModel::find($id);
+        if(is_null($country)){
+            return response()->json(["message" => "Record Not Found!!"], 404);
+        }
+        return response()->json($country, 200);
     }
 
+    /**
+     * Method : Save in api.php
+     * This method is used to Save data in _z_country table
+     *
+    */
+
     public function countrySave(Request $request){
+
+
+        $rules = [
+            'name' => 'required|min:3',
+            'iso'  => 'required|min:2|max:3'
+        ];
+
+        $validator = Validator::make($request->all(), $rules);
+        if($validator->fails()){
+            return response()->json($validator->errors(), 400);
+        }
+
         $country = CountryModel::create($request->all());
         return response()->json($country, 201);
     }
@@ -32,7 +58,11 @@ class CountryController extends Controller
      *
     */
 
-    public function countryUpdate(Request $request, CountryModel $country){
+    public function countryUpdate(Request $request, $id){
+        $country = CountryModel::find($id);
+        if(is_null($country)){
+            return response()->json(["message" => "Record Not Found!!"], 404);
+        }
         $country->update($request->all());
         return response()->json($country, 200);
     }
@@ -43,7 +73,11 @@ class CountryController extends Controller
      *
     */
 
-    public function countryDelete(Request $request, CountryModel $country){
+    public function countryDelete(Request $request, $id){
+        $country = CountryModel::find($id);
+        if(is_null($country)){
+            return response()->json(["message" => "Record Not Found!!"], 404);
+        }
         $country->delete();
         return response()->json(null, 204);
     }
